move tui.rs unit test -> tui_test.rs
